Use state table to find early_births_pct instead of calculating
Can’t weight the average the way I was doing because it’s a percent (so
it doesn’t make sense to divide by the total sample). I have no idea
why the totals for the other fields changed slightly when I stopped
joining the data (it wasn’t updated).